move `adapter` into `mod.rs`
