+++ crates/bevy_ecs/src/system/system_piping.rs
@@ -63,215 +63,3 @@ where
         b(value)
     }
 }
-
-/// A collection of common adapters for [piping](super::PipeSystem) the result of a system.
-pub mod adapter {
-    use crate::system::In;
-    use bevy_utils::tracing;
-    use std::fmt::Debug;
-
-    /// Converts a regular function into a system adapter.
-    ///
-    /// # Examples
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    ///
-    /// fn return1() -> u64 { 1 }
-    ///
-    /// return1
-    ///     .pipe(system_adapter::new(u32::try_from))
-    ///     .pipe(system_adapter::unwrap)
-    ///     .pipe(print);
-    ///
-    /// fn print(In(x): In<impl std::fmt::Debug>) {
-    ///     println!("{x:?}");
-    /// }
-    /// ```
-    pub fn new<T, U>(mut f: impl FnMut(T) -> U) -> impl FnMut(In<T>) -> U {
-        move |In(x)| f(x)
-    }
-
-    /// System adapter that unwraps the `Ok` variant of a [`Result`].
-    /// This is useful for fallible systems that should panic in the case of an error.
-    ///
-    /// There is no equivalent adapter for [`Option`]. Instead, it's best to provide
-    /// an error message and convert to a `Result` using `ok_or{_else}`.
-    ///
-    /// # Examples
-    ///
-    /// Panicking on error
-    ///
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    ///
-    /// // Building a new schedule/app...
-    /// let mut sched = Schedule::default();
-    /// sched.add_systems(
-    ///         // Panic if the load system returns an error.
-    ///         load_save_system.pipe(system_adapter::unwrap)
-    ///     )
-    ///     // ...
-    /// #   ;
-    /// # let mut world = World::new();
-    /// # sched.run(&mut world);
-    ///
-    /// // A system which may fail irreparably.
-    /// fn load_save_system() -> Result<(), std::io::Error> {
-    ///     let save_file = open_file("my_save.json")?;
-    ///     dbg!(save_file);
-    ///     Ok(())
-    /// }
-    /// # fn open_file(name: &str) -> Result<&'static str, std::io::Error>
-    /// # { Ok("hello world") }
-    /// ```
-    pub fn unwrap<T, E: Debug>(In(res): In<Result<T, E>>) -> T {
-        res.unwrap()
-    }
-
-    /// System adapter that utilizes the [`bevy_utils::tracing::info!`] macro to print system information.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    ///
-    /// // Building a new schedule/app...
-    /// let mut sched = Schedule::default();
-    /// sched.add_systems(
-    ///         // Prints system information.
-    ///         data_pipe_system.pipe(system_adapter::info)
-    ///     )
-    ///     // ...
-    /// #   ;
-    /// # let mut world = World::new();
-    /// # sched.run(&mut world);
-    ///
-    /// // A system that returns a String output.
-    /// fn data_pipe_system() -> String {
-    ///     "42".to_string()
-    /// }
-    /// ```
-    pub fn info<T: Debug>(In(data): In<T>) {
-        tracing::info!("{:?}", data);
-    }
-
-    /// System adapter that utilizes the [`bevy_utils::tracing::debug!`] macro to print the output of a system.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    ///
-    /// // Building a new schedule/app...
-    /// let mut sched = Schedule::default();
-    /// sched.add_systems(
-    ///         // Prints debug data from system.
-    ///         parse_message_system.pipe(system_adapter::dbg)
-    ///     )
-    ///     // ...
-    /// #   ;
-    /// # let mut world = World::new();
-    /// # sched.run(&mut world);
-    ///
-    /// // A system that returns a Result<usize, String> output.
-    /// fn parse_message_system() -> Result<usize, std::num::ParseIntError> {
-    ///     Ok("42".parse()?)
-    /// }
-    /// ```
-    pub fn dbg<T: Debug>(In(data): In<T>) {
-        tracing::debug!("{:?}", data);
-    }
-
-    /// System adapter that utilizes the [`bevy_utils::tracing::warn!`] macro to print the output of a system.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    ///
-    /// // Building a new schedule/app...
-    /// # let mut sched = Schedule::default();
-    /// sched.add_systems(
-    ///         // Prints system warning if system returns an error.
-    ///         warning_pipe_system.pipe(system_adapter::warn)
-    ///     )
-    ///     // ...
-    /// #   ;
-    /// # let mut world = World::new();
-    /// # sched.run(&mut world);
-    ///
-    /// // A system that returns a Result<(), String> output.
-    /// fn warning_pipe_system() -> Result<(), String> {
-    ///     Err("Got to rusty?".to_string())
-    /// }
-    /// ```
-    pub fn warn<E: Debug>(In(res): In<Result<(), E>>) {
-        if let Err(warn) = res {
-            tracing::warn!("{:?}", warn);
-        }
-    }
-
-    /// System adapter that utilizes the [`bevy_utils::tracing::error!`] macro to print the output of a system.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    /// // Building a new schedule/app...
-    /// let mut sched = Schedule::default();
-    /// sched.add_systems(
-    ///         // Prints system error if system fails.
-    ///         parse_error_message_system.pipe(system_adapter::error)
-    ///     )
-    ///     // ...
-    /// #   ;
-    /// # let mut world = World::new();
-    /// # sched.run(&mut world);
-    ///
-    /// // A system that returns a Result<())> output.
-    /// fn parse_error_message_system() -> Result<(), String> {
-    ///    Err("Some error".to_owned())
-    /// }
-    /// ```
-    pub fn error<E: Debug>(In(res): In<Result<(), E>>) {
-        if let Err(error) = res {
-            tracing::error!("{:?}", error);
-        }
-    }
-
-    /// System adapter that ignores the output of the previous system in a pipe.
-    /// This is useful for fallible systems that should simply return early in case of an `Err`/`None`.
-    ///
-    /// # Examples
-    ///
-    /// Returning early
-    ///
-    /// ```
-    /// use bevy_ecs::prelude::*;
-    ///
-    /// // Marker component for an enemy entity.
-    /// #[derive(Component)]
-    /// struct Monster;
-    ///
-    /// // Building a new schedule/app...
-    /// # let mut sched = Schedule::default(); sched
-    ///     .add_systems(
-    ///         // If the system fails, just move on and try again next frame.
-    ///         fallible_system.pipe(system_adapter::ignore)
-    ///     )
-    ///     // ...
-    /// #   ;
-    /// # let mut world = World::new();
-    /// # sched.run(&mut world);
-    ///
-    /// // A system which may return early. It's more convenient to use the `?` operator for this.
-    /// fn fallible_system(
-    ///     q: Query<Entity, With<Monster>>
-    /// ) -> Option<()> {
-    ///     let monster_id = q.iter().next()?;
-    ///     println!("Monster entity is {monster_id:?}");
-    ///     Some(())
-    /// }
-    /// ```
-    pub fn ignore<T>(In(_): In<T>) {}
-}
